ohrid3 Modern
The house of the wealthy Robevi family.
Before 1912, Ohrid was a township center bounded to Monastir sanjak in Manastir Vilayet (present-day Bitola). The city remained under Ottoman rule until 29 November 1912, when the Serbian army took control of the city during the Balkan Wars and later made it the capital of Ohrid district. In Ohrid, Serbian forces killed 150 Bulgarians and 500 people consisting of Albanians and Turks.[39] In September 1913 local Albanian and pro-Bulgarian Internal Macedonian Revolutionary Organization leaders rebelled against the Kingdom of Serbia. It was occupied by Kingdom of Bulgaria between 1915 and 1918 during World War I.

During Kingdom of Yugoslavia Ohrid continued to be as an independent district (Охридског округа) (1918-1922), then it became a part of Bitola Oblast (1920-1929), and then from 1929 to 1941, Ohrid was part of the Vardar Banovina. It was occupied again by Bulgaria between 1941 and 1944 during World War II. Since the days of SFR Yugoslavia Ohrid has been the municipal seat of Municipality of Ohrid (Општина Охрид). Since 1991 the town was part of the Republic of Macedonia (now North Macedonia).

On 20 November 1993, Avioimpex Flight 110 crashed near Ohrid, killing all 116 people on board. It is the deadliest aviation disaster to occur in North Macedonia.[40]